requirements.txt
Flask==2.3.3
psutil==5.9.5
gunicorn
flask-sqlalchemy
flask-login
flask-session
python-dotenv